Add files via upload
3. Australia - Python Meets the Circuit.

A Data-Driven Journey through the 2023 Formula 1 Season

In this tutorial, we are going to use the Ergast database described in the previous article, but now we will integrate all 14 CSV files into a single database by building the entities and relationships. All this data will be integrated into a unique dataset (Pandas) that will be used to filter the data for this specific Grand Prix (Australian GP) and to understand the performance of some specific drivers in this GP across different seasons. This main Python script, written in a Google Colab Notebook, will be available for download.

Round 03 | Australia 2023 | FORMULA 1 ROLEX AUSTRALIAN GRAND PRIX 2023